geoserver-lb - make GS' livenessProbe configureable
By default, the liveness probe for the geoserver slave deployments is
configured to hit a WMTS GetCapabilities. On platform containing a lot
of data, this can lead to produce huge XML documents (e.g. 5.5MB on
DataGrandEst), which is not relevant for a healthcheck.

This commit aims to make it configureable so that we can use more
convenient endpoint to test liveness.